# routes/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Music Stram descritption",
        "version": "1.0.0",
        "title": "Music Stream",
        "termsOfService": "",
        "contact": {
            "email": "hspharwinder@gmail.com"
        }
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {  
...
        "/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create user",
                "description": "Any new user can register.",
                "operationId": "createUser",
                "produces": [                    
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "name",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "password",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "email",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "type",
                        "description": "type-2 for user",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "enum": [2],
                            "default": 2
                        }                       
                    },
                    {
                        "in": "formData",
                        "name": "image",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "status",
                        "description": "status 1 active(aprroved), 2 not approved(in-actve)",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "enum": [
                                1,
                                2
                            ],
                            "default": 2
                        }
                    },
                    {
                        "in": "formData",
                        "name": "phone_no",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "description",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created user, responsed json Array",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Fail to create user"
                    },
                    "409": {
                        "description": "Duplicate entry"
                    }
                }
            }
        },     
...
        "/createartist": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create Artist",
                "description": "Only artist can register",
                "operationId": "createArtist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "name",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "email",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "type",
                        "description": "Type-3 for artist",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                2
                            ],
                            "default": 2
                        }
                    },
                    {
                        "in": "formData",
                        "name": "image",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "status",
                        "description": "Status 1(acitve/approved), Status 2(in-active/not approved)",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                1,
                                2
                            ],
                            "default": 2
                        }
                    },
                    {
                        "in": "formData",
                        "name": "phone_no",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "description",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {       
                        "description": "Successfully created artist, responsed json Array",                     
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    },
                    "400": {
                        "description": "Fail to create Artist"
                    },
                    "409": {
                        "description": "Duplicate entry"
                    }
                }
            }
        },      
...
        "/allVideosWithArtistId": {
            "post": {
                "tags": [
                    "Song"
                ],
                "summary": "Get all video song, pass artist id",
                "description": "Return all video songs with artist name by passing artist id                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             ",
                "operationId": "allVideosWithArtistId",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "artistId",
                        "description": "Pass artist id",
                        "required": true,
                        "type": "integer"                        
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, responsed json Array",
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    },
                    "400": {
                        "description": "Fail to retreive record"
                    }
                }
            }
        },
        "/allArtist": {
            "get": {
                "tags": [
                    "Song"
                ],
                "summary": "Get all artist with thier number of songs ",
                "description": "Return all artist with thier number of songs                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               ",
                "operationId": "allArtist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "responses": {
                    "200": {
                        "description": "Success, responsed json Array",
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    },
                    "400": {
                        "description": "Fail to retreive record"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {  
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "type": {
                    "type": "integer",
                    "format": "int(11)"
                },
                "image": {
                    "type": "string",
                    "description": "First upload your image through 'filePost' endpoints and paste here responed image path"                   
                },
                "status": {
                    "type": "integer",
                    "format": "tinyInt"
                    
                },
                "phone_no": {
                    "type": "string",
                    "format": "mobile"                    
                },
                "description": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
        "Artist": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "type": {
                    "type": "integer",
                    "format": "int(11)"
                },
                "image": {
                    "type": "string",
                    "description": "First upload your image through 'filePost' endpoints and paste here responed image path"
                },
                "status": {
                    "type": "integer",
                    "format": "tinyInt"
                },
                "phone_no": {
                    "type": "string",
                    "format": "mobile"
                },
                "description": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Artist"
            }
        },
        "Song":{
            "type": "object",
            "properties": {
                "name":{
                    "type": "string",
                    "required":true
                }, 
                "artistId":{
                    "type": "integer",
                    "required":true
                }, 
                "type":{
                    "type": "integer",
                    "required": true
                }, 
                "filePath":{
                    "type": "string",
                    "required":true
                }, 
                "thumbnailPath":{
                    "type": "string"
                }
            },
            "xml": {
                "name": "Song"
            }
           
        },
...
    },
...
}